My Week's Work
dude i want to explode. Okay so I've updated the requirements list, and I've tried two different directions to this Morgan Freeman TTS program. Morgan2.py is hard stuck on an issue I've tried to figure out for the last two hours - the shape of the tensors can't be processed for SOME reason I haven't figured out. I think I got further in morgan2.py but I'm seriously hard-stuck. morgan.py I believe I'm less close with, but there's more available progress to be made with it. To be honest, I relied a LOT on ChatGPT and I even explored DeepSeek to try and see what I was constantly doing wrong. What you guys need to know before running this program is that you not only have to have the required libraries installed (refer to requirements.txt) when you run these programs, but you may have to either make a copy of the code for yourself to change or change the code itself (without committing the change) because in both morgan.py and morgan2.py, the train_json_path is strictly declared as a path unique to me. Even though both programs are incomplete, not replacing my filepath with your own filepath will prevent you from running the program as I left them off. I didn't get around to fixing this part to natively work for any user, I'm sorry. Replacing it with your own filepath of the metadata_train.json file should work as a temporary fix that will allow you to execute both programs with one less error. ALSO, while I was making morgan2.py, a folder called fine_tuned_speecht5 with a bunch of files inside it were automatically created. It may also happen to anyone who runs morgan2.py. I didn't explore it much, but it was too big to commit to GitHub without going out of my way to so I left it out of the commit for now, even though it may show up later. I'll think about it later. Regardless, I hope and believe this is sufficient to show as our progress.