new update for morgan.py
Fixed the biggest issues - mel spectrograms are now matching sizes and the program now runs successfully and even generates a wav file with audio! Additionally, I believe I have made this code operational for MacOS users as well as Windows users. NOTE: To adjust how much time to test this program takes, adjust the epoch count.

Next hurdle: we need to figure out how to make the audio actually sound good. I'm at a bit of a loss for right now. I need that loss value to go down as the program runs but as of right now it just keeps oscillating.